implement mechanism to sync read books from achievements
This isn't actually wired up live, yet, but it does work, correctly
generating the book data it should, and passing it through so it
saves the way it should.

Unfortunately, it is just a little too slow to run on load directly,
being around 10ms without checking for matches, and the frameskip is
definitely noticable.

So, next step, integrate libasync, and perform the checks in the background,
politely allowing frames to run as expected.  That should be fun...