run a sync from lore asynchronously on login
This implements, based on libAsync, a background scan of all the
lore book records, so we can add the ones that we have seen, as
are recorded there, but which are not present in the internal
saved variables.

While this doesn't cover all books, it does cover more than we
ever had before, so fills out gaps in history.  Quite large gaps,
in my case.

This runs every login, because why not.