automatically download vendored libraries when tagged
The `X-VendorAddon: <name> <esoui id>` tag can be used to specify
that an addon should be downloaded and extracted when building the
release package.

This helps avoid manually updating remote packages when we need to
update them, and works smoothly with the model of shipping a version
that is seen by ESO, not manually vendored internally.